+++ shared/messaging/events.go
@@ -13,3 +13,12 @@ const (
 	// Utility
 	TxGossipMessageContentType = "utility.TxGossipMessage"
 )
+
+// Helper logger for state sync tranition events
+func TransitionEventToMap(stateSyncMsg *StateMachineTransitionEvent) map[string]any {
+	return map[string]any{
+		"state_machine_event": stateSyncMsg.Event,
+		"previous_state":      stateSyncMsg.PreviousState,
+		"new_state":           stateSyncMsg.NewState,
+	}
+}