+++ consensus/doc/CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.37] - 2023-03-19
+## [0.0.0.38] - 2023-03-20
 
 - Add `fsm_handler.go` to handle FSM transition events in consensus module
 - Update State Machine mock in `utils_test.go`
 - Update state_sync module with additional function definitions 
+
+## [0.0.0.37] - 2023-03-20
+
+- Add helper function `logHelper()` for logging
 
 ## [0.0.0.36] - 2023-02-28